+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # NEXT RELEASE
+
+### v0.74.3
 * Create config.locale.messages for potentially user facing string configuration
 * Reset trendingData when reset() is called
 * Refactor out FediverseCacheKey enum